requirements.txt
imageio
numpy
gradio
moviepy
pillow